TAP driver now passes signing tests on Vista x64.
Added new settings to settings.in to better control
build process.

Removed some unneeded JYFIXMEs from source code.


git-svn-id: http://svn.openvpn.net/projects/openvpn/branches/BETA21/openvpn@1874 e7ae566f-a301-0410-adde-c780ea21d3b5

+++ tap-win32/error.c
@@ -96,7 +96,7 @@ MyDebugPrint (const unsigned char* format, ...)
 	      va_end (args);
 
 	      if (status == STATUS_SUCCESS)
-		g_Debug.out = end - g_Debug.text;
+		g_Debug.out = (unsigned int) (end - g_Debug.text);
 	      else
 		g_Debug.error = TRUE;
 	    }
